feat(auth): split in submodule for bundle optimization: - @igo2/auth for the base logic of the authentification - @igo2/auth/form for the auth form BREAKING CHANGE - Delete protected.directive does not seem to be used in our project BREAKING CHANGE - Rename AuthModule for AuthFormModule BREAKING CHANGE
