+++ packages/auth/form/ng-package.json
@@ -1,0 +1,1 @@
+{}
