+++ packages/auth/form/src/auth-facebook/auth-facebook.component.scss
