+++ packages/auth/form/src/auth-form.theme.scss
@@ -1,0 +1,9 @@
+@use './auth-form/auth-intern.theme.scss' as auth-intern;
+
+@mixin themes($theme) {
+  @include auth-intern.theme($theme);
+}
+
+@mixin colors($theme) {
+  @include auth-intern.color($theme);
+}
