+++ packages/auth/form/src/auth-form/auth-form.component.html
@@ -12,21 +12,25 @@
       (login)="onLogin()"
     >
     </igo-auth-google>
+
     <igo-auth-microsoft
       *ngIf="options.microsoft && options.microsoft.enabled !== false"
       (login)="onLogin()"
     >
     </igo-auth-microsoft>
+
     <igo-auth-microsoftb2c
       *ngIf="options.microsoftb2c && options.microsoftb2c.enabled !== false"
       (login)="onLogin()"
     >
     </igo-auth-microsoftb2c>
+
     <igo-auth-facebook
       *ngIf="options.facebook && options.facebook.enabled !== false"
       (login)="onLogin()"
     >
     </igo-auth-facebook>
+
     <igo-auth-intern
       *ngIf="!options.intern || options.intern.enabled !== false"
       [allowAnonymous]="options.allowAnonymous"
