+++ packages/auth/form/src/auth-form/auth-form.component.scss
