+++ packages/auth/form/src/auth-google/auth-google.component.ts
@@ -6,11 +6,11 @@ import {
   Output
 } from '@angular/core';
 
+import { AuthService } from '@igo2/auth';
 import { ConfigService } from '@igo2/core/config';
 import { LanguageService } from '@igo2/core/language';
 
-import { AuthGoogleOptions } from '../shared/auth.interface';
-import { AuthService } from '../shared/auth.service';
+import { AuthGoogleOptions } from '../shared/auth-form.interface';
 
 @Component({
   selector: 'igo-auth-google',
