+++ packages/auth/form/src/auth-intern/auth-intern.component.ts
@@ -18,11 +18,10 @@ import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatInputModule } from '@angular/material/input';
 import { MatProgressSpinner } from '@angular/material/progress-spinner';
 
+import { AuthService } from '@igo2/auth';
 import { LanguageService } from '@igo2/core/language';
 
 import { TranslateModule } from '@ngx-translate/core';
-
-import { AuthService } from '../shared/auth.service';
 
 @Component({
   selector: 'igo-auth-intern',
