+++ packages/auth/form/src/auth-microsoft/auth-microsoft.component.scss
