+++ packages/auth/form/src/auth-microsoftb2c/auth-microsoftb2c.component.html
