+++ packages/auth/form/src/shared/index.ts
@@ -1,0 +1,3 @@
+export * from './auth-form.interface';
+export * from './auth-storage.interface';
+export * from './auth-storage.service';
