+++ packages/auth/package.json
@@ -28,6 +28,12 @@
   "exports": {
     ".": {
       "import": "./src/public_api.ts"
+    },
+    "./form": {
+      "import": "./form/src/public_api.ts"
+    },
+    "./environment": {
+      "import": "./environment/src/public_api.ts"
     }
   },
   "dependencies": {
