--- packages/auth/src/lib/environment/environment.interface.ts
@@ -1,6 +1,0 @@
-import { AuthOptions, AuthStorageOptions } from '../shared';
-
-export interface EnvironmentOptions {
-  auth?: AuthOptions;
-  storage?: AuthStorageOptions;
-}
