+++ packages/auth/src/lib/shared/index.ts
@@ -3,11 +3,7 @@ export * from './token.interface';
 export * from './auth.service';
 export * from './auth.interface';
 export * from './auth.interceptor';
-export * from './auth-microsoft.provider';
 export * from './logged.guard';
 export * from './auth.guard';
 export * from './admin.guard';
 export * from './profils.guard';
-export * from './protected.directive';
-export * from './auth-storage.interface';
-export * from './auth-storage.service';
