+++ packages/auth/src/public_api.ts
@@ -1,9 +1,6 @@
 /*
  * Public API Surface of auth
  */
-
-export { AuthFormComponent } from './lib/auth-form';
-export * from './lib/environment';
 export * from './lib/shared/auth.service';
 export * from './lib/shared/logged.guard';
 export * from './lib/shared/auth.guard';
@@ -11,12 +8,7 @@ export * from './lib/shared/admin.guard';
 export * from './lib/shared/profils.guard';
 export * from './lib/shared/auth.interceptor';
 export * from './lib/shared/auth.interface';
-export * from './lib/shared/auth-microsoft.provider';
-export * from './lib/shared/protected.directive';
 export * from './lib/shared/token.interface';
 export * from './lib/shared/token.service';
-export * from './lib/shared/auth-storage.interface';
-export * from './lib/shared/auth-storage.service';
 export * from './lib/auth-monitoring/auth-monitoring.provider';
 export * from './lib/auth-monitoring/auth-monitoring.service';
-export * from './lib/auth.module';
