+++ packages/integration/src/lib/environment/environment.interface.ts
@@ -1,4 +1,4 @@
-import { EnvironmentOptions as AuthEnvironmentOptions } from '@igo2/auth';
+import { AuthEnvironmentOptions } from '@igo2/auth/environment';
 import { EnvironmentOptions as CommonEnvironmentOptions } from '@igo2/common';
 import { EnvironmentOptions as ContextEnvironmentOptions } from '@igo2/context';
 import {
