+++ tsconfig.json
@@ -16,14 +16,15 @@
     "target": "ES2022",
     "lib": ["ES2022", "dom"],
     "paths": {
-      "@igo2/utils": ["packages/utils/src/public_api"],
       "@igo2/core": ["packages/core/src/public_api"],
       "@igo2/core/*": ["packages/core/*"],
+      "@igo2/auth": ["packages/auth/src"],
+      "@igo2/auth/*": ["packages/auth/*"],
       "@igo2/common": ["packages/common/src/public_api"],
-      "@igo2/auth": ["packages/auth/src/public_api"],
-      "@igo2/geo": ["packages/geo/src/public_api"],
       "@igo2/context": ["packages/context/src/public_api"],
-      "@igo2/integration": ["packages/integration/src/public_api"]
+      "@igo2/geo": ["packages/geo/src/public_api"],
+      "@igo2/integration": ["packages/integration/src/public_api"],
+      "@igo2/utils": ["packages/utils/src/public_api"]
     }
   },
   "exclude": ["dist", "scripts", "projects/**/e2e/**/*"],
